Move tests to match location
